+++ src/temperature.c
@@ -1,3 +1,3 @@
 const float NINE_FIFTHS = (float)(9.0/5.0);
 const float FIVE_NINETHS = (float)(5.0/9.0);
-const int THIRTY_TWO = 32;
+const float THIRTY_TWO = 32;
